Add separate library VatesAlgorithms.
