+++ Vates/CMakeLists.txt
@@ -11,6 +11,7 @@ if( ParaView_FOUND AND USE_PARAVIEW )
   message( STATUS "Using ParaView ${PARAVIEW_VERSION_FULL}" )
   include( ${PARAVIEW_USE_FILE} )
   add_subdirectory( VatesAPI )
+  add_subdirectory( VatesAlgorithms )
   add_subdirectory( ParaviewPlugins )
   add_subdirectory( VatesSimpleGui )
 else( ParaView_FOUND AND USE_PARAVIEW )
